requirements.txt
# dev requirements, prod requirements are found in setup.py
#

pyyaml
requests
frigg-coverage

# testing tools
pytest
pytest-cov
mock
responses
